Base: Renamed surface extraction processor sourcefiles
