+++ modules/base/src/processors/surfaceextractionprocessor.cpp
@@ -27,7 +27,7 @@
  *
  *********************************************************************************/
 
-#include <modules/base/processors/surfaceextraction.h>
+#include <modules/base/processors/surfaceextractionprocessor.h>
 
 #include <inviwo/core/properties/propertysemantics.h>
 #include <modules/base/algorithm/volume/marchingtetrahedron.h>
